changes in project structure. best model saved with pickle. gitignore update.
