+++ requirements.txt
@@ -1,5 +1,5 @@
-scikit-learn
-pandas
-numpy
-ipykernel
-pytest
+scikit-learn==1.3.2
+pandas==1.2.1
+fastapi[standard]>=0.113.0,<0.114.0
+pydantic>=2.7.0,<3.0.0
+pickleshare==0.7.5
