+++ tests/test_feature_importance.py
@@ -1,7 +1,7 @@
-from src.feature_importance import extract_feature_importance
-from src.model_training import build_full_pipeline, tune_model
-from src.data_processing import get_preprocessing_pipeline
-from src.custom_transformers import RemoveCollinearFeatures
+from app.src.feature_importance import extract_feature_importance
+from app.src.model_training import build_full_pipeline, tune_model
+from app.src.data_processing import get_preprocessing_pipeline
+from app.src.custom_transformers import RemoveCollinearFeatures
 import sys
 import os
